DOCUMENT-GENERATOR: created API microservice for bills reports
