+++ apps/admin/proxy.conf.json
@@ -1,0 +1,6 @@
+{
+  "/api": {
+    "target": "http://localhost:3333",
+    "secure": false
+  }
+}
