+++ apps/admin/src/app/core/components/model-forms/bills-model-form/bills-model-form.component.ts
@@ -9,6 +9,7 @@ import {
     Validators,
 } from '@angular/forms';
 import { ICargoTransportationBill } from '@app/api-interfaces';
+import { HttpClient } from '@angular/common/http';
 
 @Component({
     selector: 'app-bills-model-form',
@@ -58,6 +59,7 @@ export class BillsModelFormComponent {
 
     constructor(
         private readonly fb: FormBuilder,
+        private readonly http: HttpClient,
     ) {}
 
     public get servicesArray(): FormArray {
@@ -78,6 +80,8 @@ export class BillsModelFormComponent {
 
     public onSubmit() {
         console.log(this.modelFormValue);
+        this.http.post('/api/templates/bills', this.modelFormValue)
+            .subscribe(() => {});
     }
 
     public searchCustomerName(event: { query: string }) {
