+++ apps/api/document-generator/src/app/app.controller.spec.ts
@@ -15,11 +15,4 @@ describe('AppController', () => {
             providers: [AppService],
         }).compile();
     });
-
-    describe('getData', () => {
-        it('should return "Welcome to api!"', () => {
-            const appController = app.get<AppController>(AppController);
-            expect(appController.getData()).toEqual({ message: 'Welcome to api!' });
-        });
-    });
 });
