+++ apps/api/document-generator/src/app/app.service.spec.ts
@@ -12,10 +12,4 @@ describe('AppService', () => {
 
         service = app.get<AppService>(AppService);
     });
-
-    describe('getData', () => {
-        it('should return "Welcome to api!"', () => {
-            expect(service.getData()).toEqual({ message: 'Welcome to api!' });
-        });
-    });
 });
