+++ apps/api/document-generator/src/environments/environment.prod.ts
@@ -1,3 +1,3 @@
 export const environment = {
-    production: true,
+  production: true,
 };
