+++ apps/api/document-generator/src/environments/environment.ts
@@ -1,3 +1,3 @@
 export const environment = {
-    production: false,
+  production: false,
 };
