--- apps/api/jest.config.ts
@@ -1,16 +1,0 @@
-/* eslint-disable */
-export default {
-    displayName: 'api',
-    preset: '../../jest.preset.js',
-    globals: {
-        'ts-jest': {
-            tsconfig: '<rootDir>/tsconfig.spec.json',
-        },
-    },
-    testEnvironment: 'node',
-    transform: {
-        '^.+\\.[tj]s$': 'ts-jest',
-    },
-    moduleFileExtensions: ['ts', 'js', 'html'],
-    coverageDirectory: '../../coverage/apps/api',
-};
