--- apps/api/src/app/app.controller.ts
@@ -1,23 +1,0 @@
-import {
-    Controller,
-    Get,
-} from '@nestjs/common';
-
-import { Message } from '@app/api-interfaces';
-
-import { AppService } from './app.service';
-
-@Controller()
-export class AppController {
-    constructor(private readonly appService: AppService) {}
-
-    @Get('hello')
-    getData(): Message {
-        return this.appService.getData();
-    }
-
-    @Get('bye')
-    getByeData(): Message {
-        return this.appService.getData();
-    }
-}
