--- apps/api/src/app/app.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { Message } from '@app/api-interfaces';
-
-@Injectable()
-export class AppService {
-    getData(): Message {
-        return { message: 'Welcome to api!' };
-    }
-}
